Extract reusable 'FeedStoreSpecs' helper method into a shared scope so it can be use by other 'FeedStore' implementation tests
